render: generic asset rendering
