+++ manifests/bootkube/config/config-overrides.yaml
