+++ manifests/bootkube/manifests/configmap-aggregator-client-ca.yaml
@@ -5,4 +5,4 @@ metadata:
   namespace: {{ .Namespace }}
 data:
   ca-bundle.crt: |
-    {{ .AggregatorClientCA | indent 4 }}
+    {{ .Assets | load "frontproxy-ca.crt" | indent 4 }}
