+++ manifests/bootkube/manifests/configmap-kubelet-serving-ca.yaml
@@ -5,5 +5,5 @@ metadata:
   namespace: {{ .Namespace }}
 data:
   ca-bundle.crt: |
-    {{ .KubeletServingCA | indent 4 }}
+    {{ .Assets | load "ca.crt" | indent 4 }}
