+++ manifests/bootkube/manifests/configmap-sa-token-signing-certs.yaml
@@ -5,5 +5,4 @@ metadata:
   namespace: {{ .Namespace }}
 data:
   ca-bundle.crt: |
-    {{ .SATokenSigningCerts | indent 4 }}
-
+    {{ .Assets | load "serviceaccounts.public.key" | indent 4 }}
