+++ manifests/bootkube/manifests/secret-aggregator-client.yaml
@@ -5,5 +5,5 @@ metadata:
   namespace: {{ .Namespace }}
 type: SecretTypeTLS
 data:
-  tls.crt: {{ .AggregatorClientCertCrt | base64 }}
-  tls.key: {{ .AggregatorClientCertKey | base64 }}
+  tls.crt: {{ .Assets | load "openshift-aggregator.crt" | base64 }}
+  tls.key: {{ .Assets | load "openshift-aggregator.key" | base64 }}
