+++ manifests/bootkube/manifests/secret-etcd-client.yaml
@@ -5,5 +5,5 @@ metadata:
   namespace: {{ .Namespace }}
 type: SecretTypeTLS
 data:
-  tls.crt: {{ .EtcdClientCertCrt | base64 }}
-  tls.key: {{ .EtcdClientCertKey | base64 }}
+  tls.crt: {{ .Assets | load "master.etcd-client.crt" | base64 }}
+  tls.key: {{ .Assets | load "master.etcd-client.key" | base64 }}
