+++ manifests/bootkube/manifests/secret-kubelet-client.yaml
@@ -5,5 +5,5 @@ metadata:
   namespace: {{ .Namespace }}
 type: SecretTypeTLS
 data:
-  tls.crt: {{ .KubeletClientCertCrt | base64 }}
-  tls.key: {{ .KubeletClientCertKey | base64 }}
+  tls.crt: {{ .Assets | load "master.kubelet-client.crt" | base64 }}
+  tls.key: {{ .Assets | load "master.kubelet-client.key" | base64 }}
