+++ manifests/bootkube/manifests/secret-serving-cert.yaml
@@ -5,5 +5,5 @@ metadata:
   namespace: {{ .Namespace }}
 type: SecretTypeTLS
 data:
-  tls.crt: {{ .ServingCertCrt | base64 }}
-  tls.key: {{ .ServingCertKey | base64 }}
+  tls.crt: {{ .Assets | load "master.server.crt" | base64 }}
+  tls.key: {{ .Assets | load "master.server.key" | base64 }}
